# lib/api/swagger.yaml
...
swagger: '2.0'
info:
  version: "0.1"
  title: Webida Service API
  description: Webida Service API specfication
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html


# Should not include host / schemes in this spec, for server/client implementation
#  wants to override the values always, anyway.
basePath: /api
produces:
  - application/json
  - application/octet-stream
consumes:
  - application/json
paths:
...
  /wfs/{wfsId}/file/{wfsPath}:
    x-handler: handlers/wfs/file.js

    # readFile.
    get:
      tags: ["wfs"]
      description: read file data on path
      operationId: readFile
      security:
        - webida-simple-auth: []
      parameters:
        - $ref : "#/parameters/wfsId"
        - $ref : "#/parameters/wfsPath"
      responses:
        "200":
          description:  file contents. content-type is app.../octet-stream or follows file name extensions
          schema:
            type: file
        default:
          description: Error
          schema:
            $ref: "#/definitions/RestError"

    # writeFile
    put:
      tags: ["wfs"]
      description: create / update file with body data
      operationId: writeFile
      consumes:
        - multipart/form-data
      security:
        - webida-simple-auth: []
      parameters:
        - $ref : "#/parameters/wfsId"
        - $ref : "#/parameters/wfsPath"
        - $ref : "#/parameters/ensure"
        - name : data
          in : formData
          required : true
          description : file contents to write.
          type : file
      responses:
        "200":
          description: OK
          schema:
            type: file
        default:
          description: Error
          schema:
            $ref: "#/definitions/RestError"

    # rename
    post:
      tags: ["wfs"]
      description: Rename a file or directory to. This api does not remove an existing one.
      operationId: rename
      security:
        - webida-simple-auth: []
      parameters:
        - $ref : "#/parameters/wfsId"
        - $ref : "#/parameters/wfsPath"
        - $ref : "#/parameters/srcPath"
        - $ref : "#/parameters/ensure"
      responses:
        "200":
          description: OK
          schema:
            type: file
        default:
          description: Error
          schema:
            $ref: "#/definitions/RestError"
  /wfs/{wfsId}/dir/{wfsPath}:
    x-handler: handlers/wfs/dir.js
    # dirTree
    get:
      tags: ["wfs"]
      description: |
        returns a directory tree of given path, for listing dir and managing file system
      operationId: dirTree
      security:
        - webida-simple-auth: []
      parameters:
        - $ref : "#/parameters/wfsId"
        - $ref : "#/parameters/wfsPath"
        - name : maxDepth
          description : Maximum depth of tree. Set -1 to build a full tree, 0 to stat, 1 to plain list.
          in : query
          type : integer
          required: true
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DirEntry'
        default:
          description: Error
          schema:
            $ref: "#/definitions/RestError"

    # createDir
    put:
      tags: ["wfs"]
      description: create a directory at the path. will return error when wfsPath exists and not empty
      operationId: createDir
      security:
        - webida-simple-auth: []
      parameters:
        - $ref : "#/parameters/wfsId"
        - $ref : "#/parameters/wfsPath"
        - $ref : "#/parameters/ensure"
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/RestOK"
        default:
          description: Error
          schema:
            $ref: "#/definitions/RestError"

    # move
    post:
      tags: ["wfs"]
      description: move file or directory to given path. works like mv command
      operationId: move
      security:
        - webida-simple-auth: []
      parameters:
        - $ref : "#/parameters/wfsId"
        - $ref : "#/parameters/wfsPath"
        - $ref : "#/parameters/srcPath"
        - $ref : "#/parameters/removeExisting"
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/RestOK"
        default:
          description: Error
          schema:
            $ref: "#/definitions/RestError"
...
  /wfs/{wfsId}/ops/replace:
    x-handler: handlers/wfs/replace.js
    post:
      description: replace file contents with regex matching
      operationId: replace
      security:
        - webida-simple-auth: []
      parameters:
        - $ref: "#/parameters/wfsId"
        - $ref: "#/parameters/wfsPathList"
        - $ref: "#/parameters/pattern"
        - name: replaceTo
          in: query
          description: string to replace with
          type: string
          required: true
        - $ref: "#/parameters/ignoreCase"
        - $ref: "#/parameters/wholeWord"
      responses:
        "200":
          description: done
          schema:
            $ref: "#/definitions/RestOK"
        default:
          description: Error
          schema:
            $ref: "#/definitions/RestError"

  # mstat, mcopy and mmove is not defined yet. will be defined in next release of spec, like below
  # /wfs/{wfsId}/ops/mcopy/{wfsPath}: PUT
  # /wfs/{wfsId}/ops/mmove/{wfsPath}: POST
  # /wfs/{wfsId}/ops/mstats:
  #   get:
  #     tags: ["ops"]
  #     description: get stats of given path.
  #     operationId: mstats
  #     security:
  #       - webida-simple-auth: []
  #     parameters:
  #       - $ref : "#/parameters/wfsId"
  #       - $ref : "#/parameters/wfsPathList"
  #     responses:
  #       "200":
  #         description: stats list
  #         schema:
  #           type: array
  #           items:
  #             $ref: "#/definitions/Stats"
  #     default:
  #       description: Error
  #       schema:
  #         $ref: "#/definitions/RestError"
...
securityDefinitions:
  webida-simple-auth:
    type: apiKey
    name: Authorization
    in: header
    x-authorize: webida-simple-auth.js
parameters:
  wfsId:
    name: wfsId
    in: path
    description: webida file system id (same to workspace id) to access.
    required: true
    type: string
  wfsPath:
    name: wfsPath
    in: path
    description: |
      webida file system path to access. without heading /. should be placed at the end of path arguments
    required: true
    type: string
    pattern: .*$
  srcPath:
    name: srcPath
    in: query
    description: source data path of some operations, with have heading /
    required: true
    type: string
  wfsPathList:
    name: wfsPathList
    in: query
    description: array of wfsPath, with heading /  (collection format may be changed by implementation)
    required: true
    type: array
    items:
      type: string
    collectionFormat: multi
  ensure:
    name : ensure
    in : query
    required : false
    description : flag to create all parent directories to create file or dir, like mkdir -p
    type : boolean
    default: false
...
  removeExisting:
    name: removeExisting
    in: query
    description: remove any existing file/dir before writing.
    required: false
    type: boolean
    default: false
  pattern:
    name: pattern
    in: query
    description: regex pattern to match
    type: string
    required: true
  ignoreCase:
    name: ignoreCase
    in: query
    description: regex matching option to ignore case
    type: boolean
    required: false
    default: false
  wholeWord:
    name: wholeWord
    in: query
    description: regex matching option to match whole word
    type: boolean
    required: false
    default: false
...
definitions:
  RestOK:
    type: object
    properties:
      message:
        type: string
  RestError:
    type: object
    description:  Contains status text, not status code.
    # It's bound to status code, but not always. For Example,
    #   some 409 error will have code 'Invalid Argument Error', not 'Conflict Error'. Client should read
    #     message property to know what happened and why, exactly.
    properties:
      code:
        type: string
      message:
        type: string
    required:
      - message
...
  Stats:
    type: object
    description: simplified/augmented fs.Stats class - see node.js doc for all properties
    required:
      - type
    properties:
      type:
        type: string
        enum: ['DUMMY', 'FILE', 'DIRECTORY', 'BLOCK_DEVICE', 'CHARACTER_DEVICE', 'LINK', 'FIFO', 'SOCKET']
      birthtime:
        type: string
        format: date-time
      mtime:
        type: string
        format: date-time
      mode:
        type: string
      size:
        type: integer
      nlink:
        type: integer
      error:
        $ref: "#/definitions/RestError"
  DirEntry:
    type: object
    description: a directory entry (file or directory) with children that represents a (sub) tree
    required:
      - name
      - stats
      - children
    properties:
      name:
        type: string
      stats:
        $ref: "#/definitions/Stats"
      children:
        type: array
        items:
          $ref: "#/definitions/DirEntry"
...